refactor(JOE-977): progressive wiki module splits (oauth + graph)
Split http-api oauth façade into oauth-helpers and oauth-token-routes
(~322 LOC façade). Extract graph legend/search/URL controls to
graph/controls.js so client stays under the 800 LOC hard limit after
CSP data-attr work. Update module guide and audit matrix.
