+++ products/wiki/docs/development/module-guide.md
@@ -34,8 +34,10 @@ internal modules when it can be tested and reviewed independently.
 | `@openwiki/workflows` | Local/Postgres write coordination and lock metrics | `packages/workflows/src/write-coordinator.ts` |
 | `@openwiki/http-api` | HTTP option/result types | `packages/http-api/src/types.ts` |
 | `@openwiki/http-api` | Request body parsing, query helpers, redirects, CORS, and response writing | `packages/http-api/src/request.ts` |
+| `@openwiki/http-api` | OAuth route façade, token grants, pure helpers | `packages/http-api/src/oauth.ts`, `oauth-token-routes.ts`, `oauth-helpers.ts` |
 | `@openwiki/http-api` | Bounded markdown render cache for server HTML | `packages/http-api/src/markdown-cache.ts` |
 | `@openwiki/http-api` | System routes such as liveness probes | `packages/http-api/src/routes/system.ts` |
+| `@openwiki/web` | Graph legend/search/URL controls (CSP-safe styling) | `packages/web/src/client/graph/controls.js` |
 | `@openwiki/core` | Derived-store runtime record validation | `openWikiDerivedRecordFromUnknown`, `openWikiIndexedRecordJsonFromUnknown` |
 | `@openwiki/core` | Proposal path and section filtering shared by Git, SQLite, and Postgres readers | `openWikiProposalTargetPaths`, `openWikiProposalSectionIds` |
 | `@openwiki/core` | Validation report artifact parsing | `validationReportFromUnknown` |
