+++ products/wiki/tests/web.test.ts
@@ -163,29 +163,32 @@ test("web asset build writes a two-theme component preview gallery", async () =>
 
     const graphClient = await readFile(path.join(root, "assets", "graph", "index.js"), "utf8");
     assert.match(graphClient, /export function initGraphs/);
-    assert.match(graphClient, /applyGraphHeightFromDataset/);
-    assert.match(graphClient, /applyChipColorsFromDataset/);
-    assert.match(graphClient, /data-chip-color/);
-    assert.doesNotMatch(graphClient, /style="--ow-chip-color/);
+    assert.match(graphClient, /from "\.\/controls\.js"/);
     assert.match(graphClient, /ArrowLeft/);
     assert.match(graphClient, /graphScopeFromParams/);
     assert.match(graphClient, /graphScope === "neighborhood"/);
-    assert.match(graphClient, /setOptionalParam\(url, "scope"/);
-    assert.match(graphClient, /setOptionalParam\(url, "depth"/);
-    assert.match(graphClient, /setOptionalParam\(url, "types"/);
-    assert.match(graphClient, /data-openwiki-graph-chip/);
     assert.match(graphClient, /data-openwiki-graph-match/);
     assert.match(graphClient, /return 1500/);
     assert.match(graphClient, /pinnedNodeIds/);
     assert.match(graphClient, /handleGraphDoubleClick/);
     assert.match(graphClient, /expandGraphNode/);
     assert.match(graphClient, /neighborSrcTemplate/);
-    assert.match(graphClient, /updateGraphAccessibleLabel/);
-    assert.match(graphClient, /arrow keys to pan/);
     assert.match(graphClient, /Graph could not be loaded/);
     assert.match(graphClient, /aria-live/);
     assert.match(graphClient, /updateGraphNodeList/);
     assert.match(graphClient, /handleGraphNodeListKeydown/);
+    const graphControls = await readFile(path.join(root, "assets", "graph", "controls.js"), "utf8");
+    assert.match(graphControls, /applyGraphHeightFromDataset/);
+    assert.match(graphControls, /applyChipColorsFromDataset/);
+    assert.match(graphControls, /data-chip-color/);
+    assert.match(graphControls, /data-openwiki-graph-chip/);
+    assert.doesNotMatch(graphControls, /style="--ow-chip-color/);
+    assert.match(graphControls, /setOptionalParam\(url, key, value\)/);
+    assert.match(graphControls, /setOptionalParam\(url, "scope"/);
+    assert.match(graphControls, /setOptionalParam\(url, "depth"/);
+    assert.match(graphControls, /setOptionalParam\(url, "types"/);
+    assert.match(graphControls, /updateGraphAccessibleLabel/);
+    assert.match(graphControls, /arrow keys to pan/);
 
     const graphNodeListClient = await readFile(path.join(root, "assets", "graph", "node-list.js"), "utf8");
     assert.match(graphNodeListClient, /data-openwiki-graph-node-option/);
